Update utils.py
if I train molecule like Oxygen which contains only one element, the self.self_energies is a scalar and will make the operation self_energies = self.self_energies[species] crash. Besides, the species of Oxygen is [[3,3]] and is over the boundary of the energy list, so need to minus the minimum.