Fixed: symeig deprication (#627)
* Fixed: symeig deprication

This causes:
RuntimeError: This function was deprecated since version 1.9 and is now removed. Please use the `torch.linalg.eigh` function instead.

* Fix: slow tensor creation

This raises:
Creating a tensor from a list of numpy.ndarrays is extremely slow. Please consider converting the list to a single numpy.ndarray with numpy.array() before converting to a tensor.